gestion erreur basique client gestionnaire equipe

+++ src/main/java/fr/solutec/re/services/ClientService.java
@@ -2,9 +2,11 @@ package fr.solutec.re.services;
 
 import fr.solutec.re.dao.ClientDAO;
 import fr.solutec.re.entites.Client;
+import fr.solutec.re.entites.Gestionnaire;
 import fr.solutec.re.repository.ClientRepository;
 import org.springframework.stereotype.Service;
 
+import java.net.BindException;
 import java.util.*;
 
 @Service
@@ -17,13 +19,29 @@ public class ClientService {
         this.clientRepository = clientRepository;
     }
 
-    public void save(Client client){
+    public void save(Client client) throws BindException {
+        String Email = client.getEmail();
+        String password = client.getPassword();
+        Optional<Client> optionalclient = this.clientRepository.findByEmail(Email);
+        if (Email == null || password == null){
+            String message = "Veuillez remplir tout les champs obligatoires indiqués par *";
+            throw new IllegalStateException(message) ;
+        }
+        if (optionalclient.isEmpty() == false) {
+            String message = String.format ("Un compte est déja associé au mail %s", Email);
+            throw new BindException(message) ;
+
+        }
         this.clientRepository.save(client);
 
     }
 
     public Client findById(int id){
         Optional<Client> optionalclient = this.clientRepository.findById(id);
+        if (optionalclient.isEmpty()){
+            String message = String.format ("Aucun Client n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         return optionalclient.get() ;
 
     }
@@ -41,21 +59,12 @@ public class ClientService {
         return this.clientDAO.search(nparams) ;
     }
 
-    /*public void modify(Map<String, String> params){
-        Map<String, String> nparams =  new HashMap<>();
-        for (String key: params.keySet()) {
-            if(params.get(key) != null ) {
-                nparams.put(key, params.get(key) ) ;
-            }
-        }
-
-        this.clientDAO.modify(nparams) ;
-    }*/
-
-
-
-
     public void deleteById(int id){
+        Optional<Client> optionalclient = this.clientRepository.findById(id);
+        if (optionalclient.isEmpty()){
+            String message = String.format ("Aucun Client n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         this.clientRepository.deleteById(id);
 
     }
