+++ src/main/java/fr/solutec/re/services/EquipeService.java
@@ -1,12 +1,14 @@
 package fr.solutec.re.services;
 
 import fr.solutec.re.dao.EquipeDAO;
+import fr.solutec.re.entites.Client;
 import fr.solutec.re.entites.Equipe;
 
 import fr.solutec.re.enums.RoleEnum;
 import fr.solutec.re.repository.EquipeRepository;
 import org.springframework.stereotype.Service;
 
+import java.net.BindException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -22,12 +24,28 @@ public class EquipeService {
         this.equipeRepository = equipeRepository;
     }
 
-    public void save(Equipe equipe){
+    public void save(Equipe equipe) throws BindException {
+        String Email = equipe.getEmail();
+        String password = equipe.getPassword();
+        Optional<Equipe> optionalequipe = this.equipeRepository.findByEmail(Email);
+        if (Email == null || password == null){
+            String message = "Veuillez remplir tout les champs obligatoires indiqués par *";
+            throw new IllegalStateException(message) ;
+        }
+        if (optionalequipe.isEmpty() == false) {
+            String message = String.format ("Un compte est déja associé au mail %s", Email);
+            throw new BindException(message) ;
+
+        }
         this.equipeRepository.save(equipe);
     }
 
     public Equipe findById(int id){
         Optional<Equipe> optionalequipe = this.equipeRepository.findById(id);
+        if (optionalequipe.isEmpty()){
+            String message = String.format ("Aucune equipe n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         return optionalequipe.get();
 
     }
@@ -44,11 +62,21 @@ public class EquipeService {
     }
 
     public void deleteById(int id){
+        Optional<Equipe> optionalequipe = this.equipeRepository.findById(id);
+        if (optionalequipe.isEmpty()){
+            String message = String.format ("Aucune equipe n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         this.equipeRepository.deleteById(id);
 
     }
 
     public void update(int ideq, RoleEnum libelle) {
+        Optional<Equipe> optionalequipe = this.equipeRepository.findById(ideq);
+        if (optionalequipe.isEmpty()){
+            String message = String.format ("Aucune equipe n'a l'id %s", ideq);
+            throw new IllegalArgumentException(message) ;
+        }
         this.equipeDAO.update(ideq, libelle);
 
     }
