+++ src/main/java/fr/solutec/re/services/GestionnaireService.java
@@ -8,6 +8,7 @@ import fr.solutec.re.entites.Lieu;
 import fr.solutec.re.repository.GestionnaireRepository;
 import org.springframework.stereotype.Service;
 
+import java.net.BindException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -23,13 +24,30 @@ public class GestionnaireService {
         this.gestionnaireRepository = gestionnaireRepository;
     }
 
-    public void save(Gestionnaire gestionnaire){
+    public void save(Gestionnaire gestionnaire) throws BindException {
+        String Email = gestionnaire.getEmail();
+        String password = gestionnaire.getPassword();
+        Optional<Gestionnaire> optionalgestionnaire = this.gestionnaireRepository.findByEmail(Email);
+        if (Email == null || password == null){
+            String message = "Veuillez remplir tout les champs obligatoires indiqués par *";
+            throw new IllegalStateException(message) ;
+        }
+        if (optionalgestionnaire.isEmpty() == false) {
+            String message = String.format ("Un compte est déja associé au mail %s", Email);
+            throw new BindException(message) ;
+
+        }
         this.gestionnaireRepository.save(gestionnaire);
+
 
     }
 
     public Gestionnaire findById(int id){
         Optional<Gestionnaire> optionalgestionnaire = this.gestionnaireRepository.findById(id);
+        if (optionalgestionnaire.isEmpty()){
+            String message = String.format ("Aucun gestionnaire n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         return optionalgestionnaire.get();
 
     }
@@ -46,6 +64,11 @@ public class GestionnaireService {
     }
 
     public void deleteById(int id){
+        Optional<Gestionnaire> optionalgestionnaire = this.gestionnaireRepository.findById(id);
+        if (optionalgestionnaire.isEmpty()){
+            String message = String.format ("Aucun gestionnaire n'a l'id %s", id);
+            throw new IllegalArgumentException(message) ;
+        }
         this.gestionnaireRepository.deleteById(id);
 
     }
